api/requirements.txt
alembic==1.11.1
bcrypt==3.2.2
cryptography==45.0.5
fastapi==0.109.1
httpx==0.24.1
passlib[bcrypt]==1.7.4
pydantic[email]~=2.11.7
pydantic_settings==2.10.7
pymysql==1.0.3
pytest==7.3.1
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart
requests==2.32.4
sqlalchemy==2.0.15
uvicorn==0.22.0